# Clean up some todos & some refactoring