--- a/.gitignore
+++ b/.gitignore
@@ -31,5 +31,6 @@ build
 # https://www.npmjs.org/doc/misc/npm-faq.html#should-i-check-my-node_modules-folder-into-git
 node_modules
 
-lib
 dist
+
+docs