From a conversation earlier in irc with @diasdavid I started looking into this module and started fixing some todos, like only using pull-streams and fixing #11 

This should be ready for review, things done

- [x] Ensure `stream.write` is not called on undefined stream
- [x] Add api docs for public methods
- [x] Remove usage of regular streams
- [x] Add `2 level` test 
- [x] Add simple benchmark

Ref https://github.com/diasdavid/time-cache/pull/1 which improves the baseline perf in the simple benchmark roughly 4 times.

Now using `libp2p-crypto.randomBytes` for seqno generation, speed before vs after

```
Node 7.2.1
before randomSeqno x 242,064 ops/sec ±7.69% (66 runs sampled)
after  randomSeqno x 175,310 ops/sec ±3.01% (73 runs sampled)

Chrome 
before randomSeqno x 50,142 ops/sec ±2.66% (54 runs sampled)
after  randomSeqno x 110,842 ops/sec ±2.97% (52 runs sampled)
```

A bit slower in Node.js now but in the browser a lot faster, plus one dependency less in js-ipfs.